storage-device-tile: Launch gphoto2 import through gthumb-import.desktop
GThumb comes with a specific desktop file gthumb-import.desktop which
contains the command line for starting in photo import mode. The Exec
line contains a hack to unmount gphoto2 mounts so gthumb can exclusively
lock and access the device while importing. We are now launching the
import for gphoto2 devices using this desktop file.

USB devices must not be unmounted for GThumb being able to import
photos, so in this case GThumb is started by spawning the program
directly in import mode.

PS: The next major GThumb release will use gio/gvfs which should render
this workaround unneeded.

Fixes: http://bugs.meego.com/show_bug.cgi?id=61